Large amount of changes
Arduino functions were overhauled to have efficient and stable system for reading data and identifying if it is suitable for use. The general function of idle state works, yet the structs for Arduino's long term storage need more work. Efforts need to be made to also update the UI to have the added functionality of other branches. Beyond this, the calculation stage for the gantry movement needs to be set up, and the motions of the motors need to be written into functions of their own.